// src/app/model/training-assignment.ts
import {Time} from '@angular/common';

export class TrainingAssignment {
  trainingAssignmentName: string;
  assignmentId: number;
  businessUnit: string;
  businessUnitId: number;
  costCenter: string;
  costCenterId: number;
  trainingId: number;
  trainingName: string;
  date: Date;
  startTime: Date;
  location: string;
  assignedUserIds: bigint[];
  uuid: string;
  isSelected: boolean;
}